cumulative update for -G8 branch 2004-2023
* jambase removed
* ported to DOS386, macOS, linux64
* improved responce file support (split into lines, each <32K chars)
* added clean, this, nocare and depjam pseudo-targets
* increased MAXLINE to 512K
* facility to build jam2vs util
* using special easily parsed actions command script that recognizes:
  %errorlevel%, %ERRORLEVEL%
  echo, echo_cvt
  mkdir, copyfile, del, set,
  #respfile, #no-respfile
  rem, #, if, goto, not, exist
  set_filter, call_filtered, clr_filters,
  enter_critsec, leave_critsec, call, call_1core
* minor fixes to compile code with modern compilers
* prog/_jBuild for sample jam-build setup
  (and it is used to rebuild jam with jam)

+++ jam_src/expand.c
@@ -81,7 +81,7 @@ var_expand(
 	int depth;
 
 	if( DEBUG_VAREXP )
-	    printf( "expand '%.*s'\n", end - in, in );
+	    printf( "expand '%.*s'\n", (int)(end - in), in );
 
 	/* This gets alot of cases: $(<) and $(>) */
 
@@ -458,7 +458,7 @@ var_edit_parse(
 	    else
 	    {
 		fp->ptr = ++mods;
-		fp->len = strlen( mods );
+		fp->len = (int)strlen( mods );
 		mods += fp->len;
 	    }
 	}
